track config files
I can't remember whether I chose to install debian on this vm with all
generic drivers or now, but config.80, which was copied from /boot took
80 minutes.
I removed a bunch of random drivers and some filesystems for config.79
We'll see how long it takes to compile.
I know a minimal bootable kernel takes less than half an hour, but I
lost the config.